feat: opción de frecuencia "Cada 15 min" para pruebas del bot + frecuencia decimal
- Nueva opción de frecuencia 0.25h (Cada 15 min) en la receta para poder
  testear el flujo de seguimiento sin esperar horas.
- sincronizarMedicamentosDesdeModal usa parseFloat en frecuencia_horas para
  soportar sub-horas (0.25); el cron ya multiplica por 3600000 → 15 min.
- Encabezado de columna "Días" → "Duración (días)" para dejar claro que la
  frecuencia es el intervalo y los días la duración (cada Xh durante N días).

+++ public/js/receta.js
@@ -77,7 +77,7 @@ function sincronizarMedicamentosDesdeModal() {
       id: Date.now() + Math.random(),
       nombre,
       dosis: tr.querySelector('.med-dosis')?.value.trim() || '',
-      frecuencia_horas: parseInt(tr.querySelector('.med-frecuencia')?.value) || 8,
+      frecuencia_horas: parseFloat(tr.querySelector('.med-frecuencia')?.value) || 8,
       dias: parseInt(tr.querySelector('.med-dias')?.value) || 1
     };
   }).filter(Boolean);
